Initial work to create explicit userspace processes
