+++ Userland/CMakeLists.txt
@@ -1,0 +1,6 @@
+LIST(APPEND USERLAND_HEADERS sysm/sysm.hpp)
+LIST(APPEND USERLAND_SOURCES sysm/sysm_Main.cpp)
+
+ADD_LIBRARY(Userland STATIC
+	${USERLAND_HEADERS}
+	${USERLAND_SOURCES})
