+++ Userland/sysm/sysm.hpp
@@ -1,0 +1,10 @@
+#ifndef SYSM_HPP
+#define SYSM_HPP
+
+typedef enum MessageType
+{
+	MESSAGE_TYPE_REGISTER_CLIENT = 0,
+	MESSAGE_TYPE_REGISTER_SERVICE = 1,
+}MessageType;
+
+#endif
